+++ src/util/anyerror.rs
@@ -28,3 +28,34 @@ impl From<&str> for AnyError {
 }
 
 pub type Fault = Box<dyn Error>;
+
+#[derive(Debug, thiserror::Error)]
+pub struct CodeFault(pub Option<String>, pub Fault);
+
+impl From<Fault> for CodeFault {
+    fn from(value: Fault) -> Self {
+        Self(None, value)
+    }
+}
+
+impl CodeFault {
+    /// Checks if there is a source code parsing error.
+    pub fn is_source_err(&self) -> bool {
+        self.0.is_some()
+    }
+
+    /// References the source code file that produced an error, it exists.
+    pub fn as_source_file(&self) -> Option<&String> {
+        self.0.as_ref()
+    }
+
+    pub fn into_fault(self) -> Fault {
+        self.1
+    }
+}
+
+impl Display for CodeFault {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.1)
+    }
+}